rename oc10-sql driver to owncloudsql
