+++ changelog/unreleased/rename-oc10-share-manager-driver.md
@@ -1,0 +1,5 @@
+Change: rename oc10 share manager driver
+
+We aligned the oc10 SQL share manager driver name with all other owncloud spacific SQL drivers by renaming the package `pkg/share/manager/sql`  to `pkg/share/manager/owncloudsql` and changing the name from `oc10-sql` to `owncloudsql`.
+
+https://github.com/cs3org/reva/pull/2740
